Basic API functionality complete.
CRUD operations, and a very naive but functional search.

Auth is via API key, and only protects destructive operations.  Clients
are expected to speak JSON, and, if JS, to be on the same domain, but
JSONP is coming in next update.

This is about three hours of work, with about a third of that
being a chase down a rabbit hole trying to get Sinatra to let go of some
routes.

Total time: 5:30ish