Index page display proof-of-concepted.
History stuff basically works, tested on FF.

This is basically ending up as a close to functional API and a fairly
poor frontend, but it doesn't page off the end, and it maintains history
as long as the user doesn't manually edit the URL and not press enter.

About 3 or four more hours of work.  At a little over a solid workday of
effort, I think I'm going to document this up, fix a few more bugs, and
call this a submission.

9 or ten hours put into this so far.